docker.py: Do not pass main process's standard input to subprocesses
It is possible that the main process's standard input is a file or a
FIFO instead of a terminal device.  In this case, the contents of
standard input will be readable before the commands specified by the
user are executed in the Docker container.  If any subprocesses started
by this program consumes standard input before the Docker container is
started, then there will be no commands to read in the command execution
stage.  Thus, those subprocesses should not read anything from standard
input.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>