docker.py: Use `EMERGE_DEFAULT_OPTS` to set `emerge` options
The /usr/local/bin/emerge trick works, but it might be perplexing to
people unfamiliar with this implementation detail when they debug in a
container.  EMERGE_DEFAULT_OPTS in make.conf(5) is a mechanism provided
by Portage for setting default command-line options for emerge(1), which
is more preferable and transparent.

Signed-off-by: Yuan Liao <liaoyuan@gmail.com>